@@ -12,101 +12,55 @@ class Track extends File
 
     /**
      * Type of file to differentiate implementations in Fluid templates
-     *
-     * @var string
      */
-    protected $type = 'Track';
+    protected string $type = 'Track';
 
-    /**
-     * Default flag for track
-     *
-     * @var int
-     */
-    protected $default = 0;
+    protected int $default = 0;
 
-    /**
-     * Kind of track
-     *
-     * @var string
-     */
-    protected $kind = 'subtitles';
+    protected string $kind = 'subtitles';
 
-    /**
-     * Label for the track
-     *
-     * @var string
-     */
-    protected $label = '';
+    protected string $label = '';
 
-    /**
-     * Language of track
-     *
-     * @var string
-     */
-    protected $srclang = 'en';
+    protected string $srclang = 'en';
 
-    /**
-     * @return int
-     */
     public function getDefault(): int
     {
         return $this->default;
     }
 
-    /**
-     * @param int $default
-     */
     public function setDefault(int $default): self
     {
         $this->default = $default;
         return $this;
     }
 
-    /**
-     * @return string
-     */
     public function getKind(): string
     {
         return $this->kind;
     }
 
-    /**
-     * @param string $kind
-     */
     public function setKind(string $kind): self
     {
         $this->kind = $kind;
         return $this;
     }
 
-    /**
-     * @return string
-     */
     public function getLabel(): string
     {
         return $this->label;
     }
 
-    /**
-     * @param string $label
-     */
     public function setLabel(string $label): self
     {
         $this->label = $label;
         return $this;
     }
 
-    /**
-     * @return string
-     */
     public function getSrclang(): string
     {
         return $this->srclang;
     }
 
-    /**
-     * @param string $srclang
-     */
     public function setSrclang(string $srclang): self
     {
         $this->srclang = $srclang;
