@@ -9,9 +9,6 @@ interface ConstructibleFromImage
 {
     /**
      * Creates an instance of the class based on the provided Image object
-     *
-     * @param Image $value
-     * @return object
      */
-    public static function fromImage(Image $value);
+    public static function fromImage(Image $value): object;
 }
