@@ -3,28 +3,31 @@ declare(strict_types=1);
 
 namespace Sitegeist\MediaComponents\ViewHelpers\Image;
 
-use TYPO3\CMS\Core\Imaging\ImageManipulation\Area;
-use TYPO3Fluid\Fluid\Core\Rendering\RenderingContextInterface;
-use TYPO3\CMS\Core\Imaging\ImageManipulation\CropVariantCollection;
-use TYPO3Fluid\Fluid\Core\ViewHelper\Traits\CompileWithContentArgumentAndRenderStatic;
+use Sitegeist\MediaComponents\Domain\Model\CropArea;
+use Sitegeist\MediaComponents\Domain\Model\ImageSource;
+use SMS\FluidComponents\Interfaces\ImageWithCropVariants;
+use SMS\FluidComponents\Interfaces\ImageWithDimensions;
+use TYPO3Fluid\Fluid\Core\ViewHelper\AbstractViewHelper;
 
-class CropVariantViewHelper extends \TYPO3Fluid\Fluid\Core\ViewHelper\AbstractViewHelper
+class CropVariantViewHelper extends AbstractViewHelper
 {
-    use CompileWithContentArgumentAndRenderStatic;
-
-    public function initializeArguments()
+    public function initializeArguments(): void
     {
-        $this->registerArgument('image', FileInterface::class, 'FAL file');
+        $this->registerArgument('image', ImageSource::class, 'Image object');
         $this->registerArgument('name', 'string', 'name of the crop variant that should be used', false, 'default');
     }
 
-    public static function renderStatic(
-        array $arguments,
-        \Closure $renderChildrenClosure,
-        RenderingContextInterface $renderingContext
-    ): Area {
-        $arguments['image'] = $arguments['image'] ?? $renderChildrenClosure();
-        $cropVariantCollection = CropVariantCollection::create((string)$arguments['image']->getProperty('crop'));
-        return $cropVariantCollection->getCropArea($arguments['name']);
+    public function render(): CropArea
+    {
+        $this->arguments['image'] ??= $this->renderChildren();
+        if (!($this->arguments['image']->getOriginalImage() instanceof ImageWithDimensions)) {
+            return $this->arguments['image'];
+        }
+
+        if ($this->arguments['image']->getOriginalImage() instanceof ImageWithCropVariants) {
+            return new CropArea($this->arguments['image']->getOriginalImage()->getCropVariant($this->arguments['name']));
+        } else {
+            return new CropArea;
+        }
     }
 }
