@@ -5,28 +5,22 @@ namespace Sitegeist\MediaComponents\ViewHelpers\Image\Modify;
 
 use Sitegeist\MediaComponents\Domain\Model\CropArea;
 use Sitegeist\MediaComponents\Domain\Model\ImageSource;
-use TYPO3Fluid\Fluid\Core\Rendering\RenderingContextInterface;
-use TYPO3Fluid\Fluid\Core\ViewHelper\Traits\CompileWithContentArgumentAndRenderStatic;
+use TYPO3Fluid\Fluid\Core\ViewHelper\AbstractViewHelper;
 
-class CropViewHelper extends \TYPO3Fluid\Fluid\Core\ViewHelper\AbstractViewHelper
+class CropViewHelper extends AbstractViewHelper
 {
-    use CompileWithContentArgumentAndRenderStatic;
-
-    public function initializeArguments()
+    public function initializeArguments(): void
     {
         $this->registerArgument('imageSource', ImageSource::class, 'Image source (if not provided via content)');
         $this->registerArgument('crop', CropArea::class, 'Crop area');
     }
 
-    public static function renderStatic(
-        array $arguments,
-        \Closure $renderChildrenClosure,
-        RenderingContextInterface $renderingContext
-    ): ImageSource {
-        $imageSource = $arguments['imageSource'] ?? $renderChildrenClosure();
-        if ($arguments['crop']) {
+    public function render(): ImageSource
+    {
+        $imageSource = $this->arguments['imageSource'] ?? $this->renderChildren();
+        if ($this->arguments['crop']) {
             $croppedSource = clone $imageSource;
-            $croppedSource->setCrop($arguments['crop']);
+            $croppedSource->setCrop($this->arguments['crop']);
         }
         return $croppedSource ?? $imageSource;
     }
