@@ -5,47 +5,42 @@ namespace Sitegeist\MediaComponents\ViewHelpers\Image;
 
 use Sitegeist\MediaComponents\Domain\Model\ImageSource;
 use Sitegeist\MediaComponents\Domain\Model\SourceSet;
-use TYPO3\CMS\Core\Utility\GeneralUtility;
-use TYPO3\CMS\Extbase\Service\ImageService;
-use TYPO3Fluid\Fluid\Core\Rendering\RenderingContextInterface;
-use TYPO3Fluid\Fluid\Core\ViewHelper\Traits\CompileWithContentArgumentAndRenderStatic;
+use SMS\FluidComponents\Interfaces\ImageWithDimensions;
+use TYPO3Fluid\Fluid\Core\ViewHelper\AbstractViewHelper;
 
-class SrcsetViewHelper extends \TYPO3Fluid\Fluid\Core\ViewHelper\AbstractViewHelper
+class SrcsetViewHelper extends AbstractViewHelper
 {
-    use CompileWithContentArgumentAndRenderStatic;
-
-    public function initializeArguments()
+    public function initializeArguments(): void
     {
         $this->registerArgument('imageSource', ImageSource::class, 'Image source (if not provided via content)');
         $this->registerArgument('srcset', SourceSet::class, 'srcset definition');
         $this->registerArgument('base', ImageSource::class, 'Base image for pixel density calculations');
     }
 
-    public static function renderStatic(
-        array $arguments,
-        \Closure $renderChildrenClosure,
-        RenderingContextInterface $renderingContext
-    ): string {
-        $arguments['imageSource'] = $arguments['imageSource'] ?? $renderChildrenClosure();
-
-        if ($arguments['srcset'] instanceof SourceSet) {
-            return self::generateSrcsetString($arguments['imageSource'], $arguments['srcset'], $arguments['base']);
-        } else {
-            return $arguments['imageSource']->getPublicUrl();
+    public function render(): string
+    {
+        if (!$this->arguments['srcset'] instanceof SourceSet) {
+            return '';
         }
+        $this->arguments['imageSource'] ??= $this->renderChildren();
+
+        if (!($this->arguments['imageSource']->getOriginalImage() instanceof ImageWithDimensions)) {
+            return '';
+        }
+
+        return self::generateSrcsetString($this->arguments['imageSource'], $this->arguments['srcset'], $this->arguments['base']);
     }
 
     public static function generateSrcsetString(ImageSource $imageSource, SourceSet $srcset, ImageSource $base = null): string
     {
         $output = [];
-        $base = $base ?? $imageSource;
+        $base ??= $imageSource;
         $widths = $srcset->getSrcsetAndWidths($base->getWidth());
-        $imageService = GeneralUtility::makeInstance(ImageService::class);
         $localImageSource = clone $imageSource;
 
         foreach ($widths as $widthDescriptor => $width) {
-            $localImageSource->setScale($width / $imageSource->getOriginalImage()->getWidth());
-            $output[] = $imageService->getImageUri($localImageSource->getImage()->getFile()) . ' ' . $widthDescriptor;
+            $localImageSource->setScale($width / $localImageSource->getCroppedWidth());
+            $output[] = (string) $localImageSource . ' ' . $widthDescriptor;
         }
 
         return implode(', ', $output);
